requirements.txt
alembic==1.13.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
psycopg2==2.9.9
pyTelegramBotAPI==4.17.0
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
python-dotenv==1.0.1